Surface apply_monkey_patches directly in the util namespace
